Move "comments" URLs into notice_comments module.
