+++ notice_comment/tests/views_tests.py
@@ -3,7 +3,7 @@ import mock
 
 from django.test import SimpleTestCase, override_settings
 
-from regulations.views import comment
+from notice_comment import views
 
 
 @override_settings(
@@ -11,6 +11,7 @@ from regulations.views import comment
     ATTACHMENT_ACCESS_KEY_ID='test-access-key',
     ATTACHMENT_SECRET_ACCESS_KEY='test-secret-key',
     ATTACHMENT_MAX_SIZE=42,
+    ROOT_URLCONF='notice_comment.urls',
 )
 class TestUploadProxy(SimpleTestCase):
 
@@ -22,7 +23,7 @@ class TestUploadProxy(SimpleTestCase):
         generate_presigned.side_effect = ['first-url', 'second-url']
         get_random.return_value = 'not-so-random'
         mock_time.return_value = 123
-        resp = self.client.get('/comments/attachment?size=42&name=foo.pdf')
+        resp = self.client.get('/attachment?size=42&name=foo.pdf')
         self.assertEqual(resp.status_code, 200)
         body = json.loads(resp.content.decode())
         generate_presigned.assert_any_call(
@@ -38,7 +39,7 @@ class TestUploadProxy(SimpleTestCase):
         generate_presigned.assert_any_call(
             ClientMethod='get_object',
             Params={
-                'ResponseExpires': 123 + comment.PREVIEW_EXPIRATION_SECONDS,
+                'ResponseExpires': 123 + views.PREVIEW_EXPIRATION_SECONDS,
                 'ResponseContentDisposition': 'attachment; filename="foo.pdf"',
                 'Bucket': 'test-bucket',
                 'Key': get_random.return_value,
@@ -49,19 +50,19 @@ class TestUploadProxy(SimpleTestCase):
         self.assertEqual(body['urls']['get'], 'second-url')
 
     def test_get_url_empty(self):
-        resp = self.client.get('/comments/attachment?size=0&name=foo.pdf')
+        resp = self.client.get('/attachment?size=0&name=foo.pdf')
         self.assertEqual(resp.status_code, 400)
         body = json.loads(resp.content.decode())
         self.assertEqual(body['message'], 'Invalid attachment size')
 
     def test_get_url_over_limit(self):
-        resp = self.client.get('/comments/attachment?size=43&name=foo.pdf')
+        resp = self.client.get('/attachment?size=43&name=foo.pdf')
         self.assertEqual(resp.status_code, 400)
         body = json.loads(resp.content.decode())
         self.assertEqual(body['message'], 'Invalid attachment size')
 
     def test_get_url_invalid_extension(self):
-        resp = self.client.get('/comments/attachment?size=42&name=foo.exe')
+        resp = self.client.get('/attachment?size=42&name=foo.exe')
         self.assertEqual(resp.status_code, 400)
         body = json.loads(resp.content.decode())
         self.assertEqual(body['message'], 'Invalid attachment type')
